Addressed review feedback from @dpaiton
